data struct key file init
data struct key file added

Next steps

1. Check integrity by reading
2, Read the data into a global variable
If the value we want is not in there? Reset

If the value we want is in there? Run And Parse the data into global_variable, OR
We run read file every single time, we want to find data related to the data we want (slower run time but faster developmental time)